requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
azure-common
azure-functions
azure-storage-blob
azure-storage-common
beautifulsoup4==4.8.0
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
cryptography==2.7
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.4.0
google-cloud-storage==1.18.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
grpcio==1.23.0
httplib2==0.13.1
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
msgpack==0.6.1
protobuf==3.9.1
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycparser==2.19
pylint==2.3.1
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
soupsieve==1.9.3
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.3
wrapt==1.11.2
https://test-files.pythonhosted.org/packages/07/28/e46d3223076e249fa4b6efae3c21d8eaddad0b8a505399e6dd69e95f14fe/sosi_configuration_manager-0.0.1565983147-py3-none-any.whl